Moved code for commands into their own files
